move endpoint to web
