+++ .env.example
@@ -69,5 +69,3 @@ TELEGRAM_TOKEN=12345678:abcdefghijklmnop
 
 TICKET_HOST=https://example.org
 TICKET_APIKEY=xxx
-
-API_YEAR_IN_REVIEW_TOKEN=
