+++ config/auth.php
@@ -102,8 +102,4 @@ return [
         ],
     ],
 
-    'api' => [
-        'year-in-review-token' => env('API_YEAR_IN_REVIEW_TOKEN'),
-    ],
-
 ];
